requirements.txt
fastapi
uvicorn
pydantic-settings
chromadb
python-multipart
python-dotenv
google-generativeai
pypdf
